pyproject.toml
[project]
name = "demo"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "assemblyai>=0.45.1",
    "gradio>=5.31.0",
    "langchain-chroma>=0.1.2",
    "langchain-huggingface>=1.0.1",
    "langchain[openai]>=1.0.0",
    "openai>=1.82.0",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.1",
    "sentence-transformers>=5.1.2",
    "simple-icd-10>=2.1.1",
    "torch>=2.7.0",
    "torchaudio>=2.7.0",
    "transformers>=4.52.3",
]
